Move TestDependencyModel to outer type
